kafka_consumer/requirements.txt
kafka-python
scikit-learn
pandas